requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-allauth==0.51.0
django-ckeditor==6.4.2
django-js-asset==2.0.0
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.16.2
whitenoise==6.2.0
